🔧Share frontmatter aliases, improve validation naming
