+++ .changeset/strange-kids-tap.md
@@ -1,0 +1,6 @@
+---
+'myst-frontmatter': patch
+'myst-config': patch
+---
+
+Share frontmatter aliases, improve validation naming
