+++ packages/myst-config/src/project/validators.ts
@@ -1,4 +1,8 @@
-import { PROJECT_FRONTMATTER_KEYS, validateProjectFrontmatterKeys } from 'myst-frontmatter';
+import {
+  FRONTMATTER_ALIASES,
+  PROJECT_FRONTMATTER_KEYS,
+  validateProjectFrontmatterKeys,
+} from 'myst-frontmatter';
 import type { ValidationOptions } from 'simple-validators';
 import {
   defined,
@@ -10,13 +14,8 @@ import {
 import type { ProjectConfig } from './types.js';
 
 const PROJECT_CONFIG_KEYS = {
-  optional: ['remote', 'index', 'exclude', 'plugins'].concat(PROJECT_FRONTMATTER_KEYS),
-  alias: {
-    jupyter: 'thebe',
-    author: 'authors',
-    affiliation: 'affiliations',
-    export: 'exports',
-  },
+  optional: ['remote', 'index', 'exclude', 'plugins', ...PROJECT_FRONTMATTER_KEYS],
+  alias: FRONTMATTER_ALIASES,
 };
 
 function validateProjectConfigKeys(value: Record<string, any>, opts: ValidationOptions) {
